Another change to the kallysto datastore to isolate inside _kallysto.
This was a fairly large change to the way in which the kallyso datastore
is managed. It is now fully contained with a _kallysto dir inside the
pub src directory mainly to isolate it more from user-generated files.

In addition this led to a significant improvement in the management of
paths needed to navigate the datastore.